Apply linter
