+++ docs/conf.py
@@ -16,7 +16,7 @@
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
 import os
-import sphinx_rtd_theme
+
 
 # -- Project information -----------------------------------------------------
