+++ setup.cfg
@@ -15,4 +15,5 @@ exclude =
     .env, 
     .venv, 
     .direnv
+extend-ignore = E203
 indent-size = 4
