--- src/elemeno_ai_sdk/cos/file_utils/__init__.py
@@ -1,1 +1,0 @@
-from .zip_utils import ZipUtils
