--- src/elemeno_ai_sdk/cos/types/__init__.py
@@ -1,1 +1,0 @@
-from .cos_object import CosObject
