+++ src/elemeno_ai_sdk/ml/features/ingest/source/base_source.py
@@ -1,5 +1,5 @@
 import abc
-from typing import Dict, List, Optional
+from typing import Optional
 
 import pandas as pd
