+++ src/elemeno_ai_sdk/ml/features/ingest/source/gcs.py
@@ -1,5 +1,4 @@
-import os
-from io import BytesIO, StringIO
+from io import BytesIO
 from typing import Optional
 
 import pandas as pd
