+++ tests/test_elastic.py
@@ -1,6 +1,6 @@
 from elemeno_ai_sdk.ml.features.feature_store import FeatureStore
 from elemeno_ai_sdk.ml.features.feature_table import FeatureTable
-from elemeno_ai_sdk.ml.features.ingest.sink.ingestion_sink_builder import IngestionSinkBuilder, IngestionSinkType
+from elemeno_ai_sdk.ml.features.ingest.sink.ingestion_sink_builder import IngestionSinkType
 from elemeno_ai_sdk.ml.features.ingest.source.elastic import ElasticIngestionSource
